Add confirmation dialog when deleting engines and projects.

+++ Launcher/ViewModels/EngineViewModel.cs
@@ -38,8 +38,15 @@ public partial class EngineViewModel(Engine engine, IFilesService files, IEngine
     }
 
     [RelayCommand]
-    private void DeleteEngine()
+    private async Task DeleteEngine()
     {
-        engineManager.DeleteEngine(engine);
+        var vm = new ConfirmDialogModel($"Are you sure you want to delete {engine.Name}?");
+        var result = await vm.ShowDialog();
+        if (result is null) return;
+
+        if (result.Result == ConfirmAction.Yes)
+        {
+            engineManager.DeleteEngine(engine);
+        }
     }
 }
